Fix CELT CBR overshoot via fixed-storage range encoder, add RFC 7845 pre-skip
RangeEncoder::try_done_sized finalizes into an exact byte budget by
mirroring libopus's fixed-buffer ec_enc_done layout, replacing the
CBR encoder's silent overshoot with a rejected-allocation error.
OggOpusEncoder now signals a 120-sample pre-skip and pads output so
granule positions survive pre-skip removal round-trip correctly.

+++ tpt-av-cadence-aac/src/sbr/mod.rs
@@ -718,9 +718,9 @@ impl Sbr {
                 if e as i8 != e_a[0] && e as i8 != e_a[1] {
                     let (phi0, mut phi1) = match indexsine {
                         0 => (1.0f32, 0.0f32),
-                        1 => (0.0, (1 - 2 * (kx & 1)) as f32),
+                        1 => (0.0, (1i32 - 2 * (kx as i32 & 1)) as f32),
                         2 => (-1.0, 0.0),
-                        _ => (0.0, -((1 - 2 * (kx & 1)) as f32)),
+                        _ => (0.0, -((1i32 - 2 * (kx as i32 & 1)) as f32)),
                     };
                     let mut noise = indexnoise;
                     for m in 0..m_max {
